Support updateAdnByIndex for USIM entity files
The EF IDs for entity files under DF telecom are not constants and
so need to be copied from existing record when updating rather than
being hardcoded.

Also add some methods to AdnRecord to get the max sizes for the name and phone number and encode and decode the name.

Bug: 154363919
Test: atest CtsSimPhonebookProviderTestCases
Change-Id: Ie2ef3cd0323e99e102de06e741807b381a77efb4