Change EuiccConnector to only refresh eSIM profile list when rebinding
EuiccConnector picks the best LPA to use. In order to detect whether
there is a new LPA installed, it listens to all app package status
changes which including deleting, installation, and modification. The
profile list refresh was placed in the wrong place which will be
triggered on any app package status changes. In order to fix this issue,
we should only refresh the eSIM profile list when EuiccConnector rebind
a LPA.
Bug: 171616491
Test: Manually tested

Change-Id: I7c49734b10ea4df7615dc9fbc6db6d2ae3169185
Merged-In: I7c49734b10ea4df7615dc9fbc6db6d2ae3169185