Prioritize to use a preferred APN for tethering if it can handle DUN
For DUN connection requesting, the list returned by fetchDunApns() is
used as waiting APNs of ApnContext. But, fetchDunApns() does not
prioritize preferred APN. It means that it may lead to user unexpected
DUN APN is connected or unnecessary DUN connection requests.
Therefore, if the preferred APN can handle the DUN type, it should be
set at the top of the list to use preferentially.

Test: manual
Test: auto - Passed DcTrackerTest
Bug: 184105189
Change-Id: I0cedb382e1aa1b806a67ca17f14a4110db4d9b7c