Fix TelephonyNetworkFactoryTest#testRequests
This test was relying on NetworkFactory being able to resend
all requests in the system on its own according to the set filters.
S moves the evaluation of the filters into the system server, but
this test had no provision for mocking the system server behavior.

Fix this by implementing a very basic fake of the system server
behavior.

Bug: 184157613
Test: atest TelephonyNetworkFactoryTest
Change-Id: I24f4f12c6579adbe395c3d4ba5cff79998d06cb7