Consider isRadioPowerOn before updating roaming state
Context:
ImsPhone keeps track of the last known roaming state i.e. mRoaming
state in order to decide which WFC mode to apply (home vs roaming).
The state is also used to decide if WFC prompt should be shown: it
should not be if WFC roaming is disallowed and is currently roaming.

Issue:
It's found that ServiceState can report IN_SERVICE when device in
airplane mode i.e. isRadioPowerOn=false and Wi-Fi connected. Hence
mRoaming is udpated to true, which may not be correct.

Fix:
Consider isRadioPowerOn=false as not IN_SERVICE, so do not update the
last known roaming state i.e. mRoaming.

Risk of regression:
Minimal. The WFC mode doesn't matter in airplane mode since voice over
cellular is not possible anyway.

Bug: 172975077
Test: fake roaming, turn on airplane mode + Wi-Fi: mRoaming not updated.
Change-Id: I26242d1904869265fcd5ea273ea49fb6fc6882d0