Remove dependency on Intent for IMS deregistration timeout event
When the device is under high load situations or booting up for the
first time, Intent delivery to the telephony process is not
guaranteed to be timely. Old code was using a PendingIntent + AlarmManager
to deliver the event to time out turning off the radio if IMS was
registered when Airplane mode was enabled, which would sometimes not
be delivered for a long time after Airplane mode was enabled.

This change removes the AlarmManager pattern and uses a more traditional
Handler approach to triggering a timeout.

Fixes: 178682618
Test: atest FrameworksTelephonyTests
Change-Id: I38fbc2f4fe24b8ad2f6721149e2b7d77fa9d590b