Allow device ImsService implementing RCS to use CarrierMessagingService
The ImsService implementing RCS needs to intercept SMS messages
during RCS provisioning as part of the authentication process.
Extend the existing mechanism for carrier ImsServices to support
this feature to also include the device ImsService.

1) Instead of using the ImsService defined in the CarrierConfig
to figure out which package telephony should bind to, use the
configured RCS ImsService package name from ImsResolver, which
already determines which ImsService to bind to for RCS features.

2) Update CarrierSmsUtils to use this information to get the
ImsService implementing RCS instead of only the carrier configured
ImsService (could be device or carrier service, depending on what
is configured/bound).

3) Update docs to clarify the difference now

Bug: 188208808
Test: atest  CtsTelephonyTestCases FrameworksTelephonyTests
Change-Id: I1594450b83ea99bc8c3c0aa76a4ea49aedeb5665