Remove onCallInitiating/Progressing order exceptions.
The onCallInitiating API should be called before onCallProgressing.
There was code present to validate that the OEM is implementing the
behavior correctly.
However unfortunately the ImsCall.Listener instance is shared among all
calls, so in multi-call scenarios the IllegalStateExceptions can trip when
there is not actually an invalid scenario.  Moving to a non-shared
listener is a bigger change and not something we can easily regression
test, so removing the exception throws is a reasonable solution.

Test: Dialing regression testing; this an edge case which is not
possible to produce manually.
Bug: 189291969

Change-Id: I075f5081a3a04d4adb6674a8deb4ae6e155ce5ad