Transition to current state on preferred network type changed
This essentially reverts ag/12980530, which was made because the network
types chosen for the preferred network types could be unsupported by the
modem and we would display an invalid network type. However, starting
from ag/13297509, preferred network types are ANDed with the modem-
capable network types, so the preferred network types will always be
valid and there is no need to wait for a second event to update the
current state.

Test: atest NetworkTypeControllerTest
Bug: 185547798
Change-Id: Icbfb0c01bc9bed61a9e31779e982b365fd11b21e
Merged-In: Icbfb0c01bc9bed61a9e31779e982b365fd11b21e