sync android DDS to modem after APM exit
This is to address the corner cases where modem DDS and android DDS are out-of-syncsometimes. in this case, after user exit apm, framework didn't issue any correction command to modem with the correct DDS while modem was doing some self-correction with a different preferred data modem selected. In the long term we should have a API to query what's the current preferred data modem to detect such out-of-sync case precisely.

Bug: 175389555
Test: Manual test with APM on/off & PhoneSwitcherTest
Change-Id: Id1e51858de3abdeacbdf9b14c1e44cacf1103933