Fix ordering of calls when initiating ECT from MMI code '4'.
Per 3GPP TS 24.629 - A.2, the signalling for an explicit call transfer
should send the REFER request on the background held call with the
foreground call specified as the destination; not the other way around.

Test: Manual test; initiated two calls and dialed MMI code '4'; verified
that the ECT takes place as expected.
Bug: 154293531
Merged-In: Ic3b7e85de8dd8c15ed3e60097adb1e0518837263
Change-Id: If2030104bb9a5f5a9edd9df046d6f695d909cf7a