Set default data roaming on SIM LOADED
If device SIM is pin locked, then on boot onCarrierConfigChanged
is called with the default carrier config rather than the carrier
specific config. To prevent this, we delay setting default data roaming
until the SIM is loaded.

Ideally isCarrierConfigApplied should be stopping readConfiguration from
being called at all in SIM locked state, however this is not happening.
That part of the issue is being addressed in b/190750288

Bug: 172316767
Test: manual
Change-Id: I6f3108b209740274242cd6e6acf81d485a1ff3d0