Reset SIM loaded flag on radio unavailable
The flag is set to true when SIMs are loaded, but not set to false when
the radio is unavailable (modem reboot or modem crash).
This causes a race condition where we think SIMs are loaded and ready
to query, but they actually aren't yet.

Test: atest MultiSimSettingControllerTest
Bug: 187184692
Change-Id: Ide8ea1d46131c8be55d7453437a3c1a6e23b4b5d
Merged-In: Ide8ea1d46131c8be55d7453437a3c1a6e23b4b5d