Enable SignalThreshold with consideration of both system and apps
With "Triggered Signal Strength reporting" feature introduced
in b/164432835, SignalThreshold enableness is not only determined
by system for sysUI policy, but also by signal update request
from privileged apps.

For specific RAN with signal measurement type, the corresponding
threshold will be enabled if system and/or app request to enable it.

The enablenss of the system threshold also determines if it should
be involved in the thresholds consolidation. Only it is enabled and
honored (either in high power mode or app explicitly request it in idle)
when it should be included.

Bug: 176961228
Test: com.android.internal.telephony.GsmCdmaPhoneTest
Test: com.android.internal.telephony.ServiceStateTrackerTest
Change-Id: Ie699e009aa99dc6d01c78192f2c0132236883c6d
Merged-In: Ie699e009aa99dc6d01c78192f2c0132236883c6d
(cherry picked from commit 008d069f93bfb71277cafdc4aa7eb322082beca0)