Fix issue where redial connection change is notified on wrong phone.
notifyRedialConnectionChanged was being raised on the ImsPhone, which is
incorrect.  The only registrations for redial connection change comes
from TelephonyConnection, which registers on the default phone (e.g.
GsmCdmaPhone).  As a consequence, when a silent redial connection change
occurs, the TelephonyConnection was not being notified and we'd end up
with a stuck call.

Test: Add unit test to verify that the silent redial connection change
event is raised on the default phone, not on the ImsPhone.
Bug: 191607537

Change-Id: I50506dadfe300c3682fe091a452b619ab5ba54f4