Speculative test fix.
For some reason tests are failing on cloud instances due to this code.
It looks like the class loader can't find the static method which was
being called on TelephonyTest.  Given that the method is pretty simple,
will try to speculatively call the log.d method directly in the hope
this will correct the issue.
Bug: 192246388
Test: This is a speculative test fix.

Change-Id: Iea2337a008277832b664384d24ae54dc5a545c35