Notify CDMA call Settings when CDMA USSD codes are unavailable
When getting CDMA Supplementary Service information, throw an
Exception if SS over CDMA is disabled, which will cause an error
to be shown to the user and the setting moved to disabled.

This is to support the case where SS over UT is available and becomes
unavailable while the user is interacting with these settings, but
there is no CDMA fallback available. Throwing an exception will
cause the UI to gray out settings that are now unavailable.

Bug: 191057045
Test: manual
Change-Id: I7c6356286a79e55b362afcdc40f157d61e55d29a