Fixed that data was created in suspended state
Fixed a regression introduced by ag/15133908. The network
was created when the device was out of service, then connectivity
service saw the data state changed from CONNECTING to SUSPENDED.
The fix is always creating the data with NOT_SUSPENDED capability,
then immediately update the suspended state after that. (In fact
we already used this mechanism in aosp/1443087)

Test: atest DataConnectionTest
Fix: 193847808
Change-Id: I4217bb1715d9fb576579512226d6a68d440fb99d