Add a new version of Phone#dial that informs the caller of the Phone
When GsmCdmaPhone#dial is called, it will internally either place
the call over CS or IMS, depending on the telephony stack conditions.
Introduce a new version of dial that includes a Consumer, which will
send an update with the Phone used to place the call. This can be
used to register call event listeners on the Phone before the call
is placed, lowering the risk of race conditions due to a callback
being called before the listeners are set.

Bug: 193805172
Test: atest TelephonyTestCases
Change-Id: Ia9cc113b7bd227b38ea56657fdd698159033907d