Update subId on sim state READY.
This CL also reverts recent fixes related to faking NOT_READY
state on SIM_REFRESH. It turns out that terminal NOT_READY is
not the correct state during profile switch. And non-terminal
is not correct at profile disable time, because otherwise real
terminal NOT_READY ends up getting ignored. Conclusion is that
NOT_READY should not be faked on SIM_REFRESH.

Earlier issues are being fixed by updating subId on READY.
In R, this was not an issue because subId was updated on LOADED
'unconditionally'. In S, with aosp/1285363 subId is not always
updated. This CL removes an if check so that it is always updated
on READY.

Test: manually verified eSIM disable and eSIM profile switch
Test: atest SubscriptionInfoUpdaterTest
Test: atest UiccControllerTest
Bug: 195092083
Change-Id: I7d6d6d9e666b3eb1fda6a4de76e42df0e85d0bf6