Update comparison of new slot status with last status.
The old comparison was based on contains() which can be incorrect
if the new state of a slot is same as the existing state of one
of the other slots. contains() returns true in that case even
though the overall slot status has changed.

Test: atest UiccControllerTest
Bug: 195622658
Merged-In: I79bc992a8cd2336cd6f0c804dd3bde077ca53dd3
Change-Id: I79bc992a8cd2336cd6f0c804dd3bde077ca53dd3