Add support for message wildcard in KEY_IMS_REASONINFO_MAPPING_STRING_ARRAY
The modem may report 1014 with "call rejected by user" when
an incoming call was terminated by the originating party
before our device answers it. For some carriers we want to
keep 1014 only when the reason message reported by the modem is
"call completed elsewhere", and remap it to 510 otherwise.
With support of message wildcard, that can be achieved by
setting these two strings in
CarrierConfigManager.KEY_IMS_REASONINFO_MAPPING_STRING_ARRAY:
- "1014|call completed elsewhere|1014"
- "1014|*|510"

In addition, we should clear the map
before adding the mappings from the new carrier config into it.

Test: atest FrameworksTelephonyTests:ImsPhoneCallTrackerTest#testReasonCodeRemap,testNumericOnlyRemap,testRemapEmergencyCallsOverWfc,testRemapWfcNotAvailable,testReasonCodeRemapCarrierConfig,testReasonCodeRemapCarrierConfigWithWildcardMessage

Bug: 208172486

Change-Id: I9bf14ec920006f58499e4e24d4ab81f677ca3c19