Configure the radio power off delay timer using a device config
In android S, we now delay the radio power off during APM and
power off for N seconds to allow for the IMS service to deregister.

This is not necessary for all devices, so instead of delaying
power off in these cases where it is not necessary, create a
device configuration that allows the devices to tune the radio power
off delay if it is required, or remove the delay entirely.

Bug: 195304068
Test: manual; atest FrameworksTelephonyTests:ServiceStateTrackerTest
Change-Id: Ia6b62c98f86444b1b598af04f21d0806b6426159
Merged-In: Ia6b62c98f86444b1b598af04f21d0806b6426159