Update patch set 6
Patch Set 6:

With DSDS If you are on lte with data on - on sim 1, both sims stay in lte.
if we switch data to sim 2, sim 1 can't stay on lte for some reason and need to drop to 3g/2g.
Again it could be a Moto issue, as the whole Moto radio is very messy.

Patch-set: 6
Reviewer: Gerrit User 23396 <23396@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>