Support dialing e-call based on dialing sim's emergency number list.
Support a carrier config for dialing an emergency call based on the
dialing sim's emergency number list.

When this config is enabled, we will only consider the dialed SIM's
emergency number list when placing an emergency call instead of the current
platform behavior of using the platform's combined list of known emergency
numbers across all SIMs.

Test: atest com.android.internal.telephony.GsmCdmaPhoneTest#testEmergencyCallAnySim
Test: atest com.android.internal.telephony.GsmCdmaPhoneTest#testNotEmergencyNumberOnDialedSim1
Test: atest com.android.internal.telephony.GsmCdmaPhoneTest#testNotEmergencyNumberOnDialedSim2
Bug: 203947159
Change-Id: I292d7f27b253f71066194e3f60bd7114e579c87c