Add support for retaining / restoring NITZ signals
Add support for retaining NITZ for short periods when a network is
declared unavailable. When a network is declared available the NITZ may
be restored for use in time / time zone detection. An initial value of 5
minutes has been chosen.

This is an effort to improve time zone detection: some devices appear to
disconnect for short periods. Because carriers do not send NITZ signals
regularly, or if a device disconnects for a short period, this may help.
It may also cause problems: this reintroduces a call to
handleNetworkUnavailable() that was removed previously and there is no
guarantee that a retained / restored NITZ is from the same network as
the one the device is now attached to.

Bug: 204415203
Bug: 161967945
Test: atest tests/telephonytests/src/com/android/internal/telephony/nitz/NitzStateMachineImplTest.java
Test: treehugger
Change-Id: Id4694975fa6eb28fc07fab2192de7703dc701c67