when radio capability is without NR, timer should not start.
There is a race condition issue. The device set the radio capability
as without NR, then telephony reset the timer. After the radio
capabilty change, the modem may report NR state in the while since
the serice state is not changed immediately. If the service state
changed from NR to LTE after telephony reset the timer, then the
telephony start the timer.
Therefore, checking the radio capability before the telephony start
the timer.

Bug: 202359605
Test: atest NetworkTypeControllerTest
Change-Id: I6c9362791ae9cced601d20aec96c22a7f57b2840