Fix phone crash by multiple result messages
As switching from the live HIDL modem to Mock AIDL modem, the
mRadioProxy would be null. As startLceService() API is called, the
result RADIO_NOT_AVAILABLE would be sent first due to mRadioProxy is
null and then REQUEST_NOT_SUPPORTED would be sent again as checking
the HIDL version. Therefore, a phone exception is occurred. Move the
HIDL version check into the condition where mRadioProxy is ready to fix
the problem.

Bug: 203241555
Test: no phone crash as switching mock modem and live modem
Change-Id: I981a4d60a8f025171bf99b0cb73d7097f82f56b6