Fixed incorrect count logic
The no_cap count is incorrect when checking the capability of
the presence notification event.
If the pidfXml has rcs_cap, mmtel_cap and no_cap,
the no_cap count must be set to zero.
This is because either rcs or mmtel cap is included in the xml

Bug: b/207321490
Test: atest RcsStatsTest

Change-Id: I867a3ef33738de1444a375083a7033e597862026
Merged-In: I867a3ef33738de1444a375083a7033e597862026