Add required flag to registerReceiver call in GsmCdmaPhone
Android T adds support to allow a runtime receiver to be registered as
not exported, but to ensure apps can take advantage of this, calls to
registerReceiver must specify a flag indicating whether the receiver
should be exported for apps targeting T+ that are registering for
non-system broadcasts. This commit adds the RECEIVER_EXPORTED
flag to the receiver in GsmCdmaPhone since this broadcast is protected
by a privileged|role permission and is expected from the Dialer.

Bug: 161145287
Test: Build
Change-Id: I6d8fa18d062411823b4e4a2b157e429ee01e8ad2