Support carrier privileges linger period in CarrierPrivilegesTracker
When SIM is gone, linger the carrier privileges from the SIM
for 30 seconds, unless a SIM with different UiccRules is inserted.
This is used to support the case for switching between the
same carrier.

The CL also support the case when SIM is READY but not fully
LOADED. In this transient state, all previously loaded rules are
not reliable because they may be updated very soon.

Bug: 169444588
Test: atest CarrierPrivilegesTrackerTest
Change-Id: Ic225eebb6286654b431c8eb1b533a524b0a2b34f
Merged-In: Ic225eebb6286654b431c8eb1b533a524b0a2b34f
(cherry picked from commit 1166a0411d21fa949643e3adf9c47cde77c25081)