Fixed incorrect data icon when call received on ESIM
The previous fix ag/16729359 attempted to fix the issue
that data icon was displayed as data disabled. The fix was
listening to IMS call state and updated the data enabled
state. The problem is that when data settings manager was
trying to register for the call event in DataSettingsManager's
constructor, IMS phone was not created yet.

Fixed by moving the registration event logic into
DataSettingsManager's event handler.

Test: Tested with PSIM + ESIM and enabled
"allowed data during calls". Data icon was shown properly during
voice call.
Fix: 213511868

Change-Id: I5ff7962ed270f449b90bf4c0d3c03390ffa9b709