update DSDS call preference logs (2/2)
User experienced a bug where the "Ask Every Time" call preference
was not working properly.   Expected: prompt the user on which SIM
they want to use for an outgoing call but Actual: no prompt.

There was not enough information in the bug report to determine the
issue, nor was the bug reproducible.

Hence, some more logging statements have been added to help pinpoint
the cause.

bug: 198123192
Test: none
Change-Id: I55407959def6307940bc23fecfc37d36ccefaa62