move the responsibility of calling updateImsServiceConfig back to the ImsPhoneCallTracker
The ImsProvisioningController has no information as to the current state
of the ImsPhonCallTracker and can not wait to send the configs until after the initial set-up from the ImsPhoneCallTracker#startListeningForCalls method.
This is causing some conditions where ImsProvisioningController calls ImsManager#updateImsSericeConfig before the ImsPhoneCallTracker is able to set up the interface properly.

Bug: 219992992
Test: atest
Change-Id: I9694cd7ce58315dce99ac7ae5df12bac34da5c5d