Allow NITZ signals > year 2037
Some Android devices shouldn't need Y2038 protection. If there were
issues with years > 2038, they are best applied in the time_detector
service across all time signals and not in this one place.

If networks were previously in the habit of sending values >= "38" for
the year, these would previously have been discarded. Now they will be
passed on to the time detector (which may then discard them).

Bug: 204193177
Test: atest tests/telephonytests/src/com/android/internal/telephony/NitzDataTest.java
Test: atest tests/telephonytests/src/com/android/internal/telephony/ServiceStateTrackerTest.java
Change-Id: Icde3b0d8797c51758449b0d838b0f34477b65afe