[MEP] switchToSubscription with portIndex is failed.
This issue has two parts.
1. Target port validation is getting failed when the device is in SS pSIM Mode and the app calls switchToSubscription with portIndex.
2. Platform is sending the wrong slotId to LPA through EuiccConnector.

In above two cases retrieving slotIndex from cardId is failed as we are using UiccCardInfo which has cardId as TelephonyManager#UNINITIALIZED_CARD_ID. As per UiccController, UiccCardInfo will have proper eSIM cardId only when the UiccCard is created. So modified logic to use UiccSlotInfo to retrieve the slotIndex.

Bug: 222680507
Test: Manually tested on sunfish device and atest FrameworksTelephonyTests:EuiccControllerTest
Change-Id: Ib5802c97a430974b039437cafc7453f19acfcd0f