fix No Call Preference bug afer switch to DSDS mode
Multiple Android users were reporting that the DSDS
mode was enabled and the settings were showing a
"No Call Preference" but calls would still default
to the sim set before the switch to DSDS mode.

Turns out that there is an inconsistency with what the
initial state of the default voice subId was in
telephony/telecom vs. what the network settings
displayed in the UI.

The fix is to set the default voice subId to
"No Preference" when the user enables DSDS mode so that
what the settings are showing and telephony/telecom are
in sync.

bug: 198123192
Test: manual, 1 new UT
Change-Id: I1e3ef517161ffeff482d7bd03db02aa5553b8acf