Enhance logs and cleanup removed connections.
Enhancing the log of the updatePhoneState logs to include the ids of
the connections instead of just a count.
Adding a new cleanupAndRemoveConnection method which is used to
remove a connection which was replaced when a new unknown connection
replaces a dialing connection (ie both on IMS).

Test: Add unit test for connection cleanup/removal.
Test: Manual call tests to verify logs.
Bug: 225961946
Change-Id: I4d9f633d5c1ebe9021cf5929a5b56f1807114e62
Merged-In: I4d9f633d5c1ebe9021cf5929a5b56f1807114e62