Honor NOT_RESTRICTED set by VCN policy
This change allows Telephony to honor the VCN's NOT_RESTRICTED policy,
for consistency with Wifi. This makes maintainance of the entire system
simpler.

In addition, this prevents apps that enumerate interfaces and networks,
binding and testing each, resulting in failures when usage of networks
is mutually exclusive (eg, some implementations of modems)

Bug: 202547696
Test: atest DataNetworkTest
Change-Id: I258ac3e203886c98bc201b1f896275893e351c5e