route dial out strings to the same number based on service state
VZ network noticed that 611 and *611 were dialing different
numbers that are meant for IVR when the phone service state
was the same(e.g ROMAING_DOMESTIC).  This is not the intended
behavior and the dial out string should be routed to the same
number remapping which is based on the service state. Dialer
handled this before but at some point stoppped.

A new carrier config bundle has been added to account for the
service state ROAMING_TYPE_INTERNATIONAL.

bug: 211116126
Test: 2 unit tests,
      manually tested by
          (1) setting the phone service state to roaming
	  (2) setting carrier config bundles
	  (3) overriding the country code (spoof location)
	  (4) placing calls and asserting the
	      correct devices were called (based on service state)
Change-Id: Ie65d837da9c169e1a9986e147222885909662344