Reduce signal strength alignment hysteresis to half of system default
QNS requires less alignment for the carrier specified thresholds.
But we have to guarantee the system threshold not being impacted firstly.
Since the alignment for application's threshold applies to both directions,
half of the system default hysteresis for alignment is good enough
to make sure system thresholds is un-impacted while introducing no
additional gaps for QNS (and other requests).

Bug: 227932356
Test: atest SignalStrengthControllerTest
Change-Id: Ife582f46208df0bd6fb0a8c1196f83a203ba4e94