Duplicate Data Selection Dialog during SUW
During SetupWizard in default SS devices we had only one active subscription,
for default DSDS devices we have 2 active Subscriptions so a dialog is being displayed to select default data during SuW as well as a full screen activity.
so during SuW we are hiding the Dialog to select default data.

Test: atest MultiSimSettingControllerTest
Test: manual test
Bug: 228130260
Change-Id: I3ede490bf08ffbb1d23f49d36e13b866ae872842