Change how reject versus missed calls are reported from ImsPhoneCallTracker.
The old logic assumed the modem would return an appropriate disconnect
cause based on missed or rejected disconnect causes reported from the IMS
stack.  Unfortunately there are some disconnect causes we were seeing in
practice which this code couldn't handle and we were mis-categorizing some
missed calls as rejected even though they were not.

Realistically we know the user's intention already, so we should just use
that to set the disconnect cause to missed or reject.

Test: Manual test on carrier network.
Test: Added new unit tests
Fixes: 228336056
Change-Id: I4c56ae2f9f52e1c0a8197ced1b26d92de33f8464