Update the SignalStrengthControllerTest UT to use valid threshold
The test cases verifies that, a NR SSRSRP value of -44 can be
a level GREAT for the default system threshold, but also can be
a level of POOR,if system threshold is overrided.

But the overrided threshold values used in this test case is
out of valid scope of SSRSRP (-140, -44) which trigger
an exception when tested with Radio HAL 2.0 UT.

To fix it, update the test SSRSRP value to be -64 and the overrided
system threshold values to be in valid scope.

All other UT cases are also checked to make sure no similar issue.

Bug: 230342465
Test: atest SignalStrengthControllerTest (with radio HAL 2.0 UT)
Merged-In: Ide60679b92c42e0225719f320544c8dd635e7bba
Change-Id: Ide60679b92c42e0225719f320544c8dd635e7bba
(cherry picked from commit 02e201316a4d87b788a8f93644d18a49b059254e)