Re-run detection under more conditions
handleAirplaneModeChanged() would only rerun time / time zone detection
if the network state (NITZ) was cleared. It should be run if the country
state was cleared too so that a proper "uncertain" suggestion will be
made in all cases.

This also adds a regression test to confirm this case is covered.

Bug: 227047106
Test: atest ./tests/telephonytests/src/com/android/internal/telephony/nitz
Change-Id: I444ce448029c09a99e41d06bb7a478958f849d15