Handle CODE_UT_OPERATION_NOT_ALLOWED.
CODE_UT_OPERATION_NOT_ALLOWED is synonymous with CODE_UT_NOT_SUPPORTED and
is returned by some carriers when attempting to query a call barring option
which is not modifiable by the user.  We want to report this to the UX
screen as a REQUEST_NOT_SUPPORTED error code so that we can specifically
prevent showing an error message in this case.

Test: Manual test; enable barring on tmobile; hard-code screen to return
operation not allowed in the case of an error; verify UX behavior is to
not show the error dialog.
Fixes: 231284401

Change-Id: I05dd203229814a2b8e92becce2e1fc707fb5bf88