Gracefully handled IP changes
Migrated the logic from DcController to DataNetwork for handling
the scenario when IP changes. Instead of tearing down the underlying
PDN, we re-create the network agent so from clients' perspective
a new network is created.

Also a minor fix on the anomaly report so we only track unwanted
network request from connectivity service.

Fix: 232348939
Test: Sanity + IWLAN handover testing on live network
      + atest DataNetworkTest (for various IP changes scenarios)
Change-Id: I27d96617358292edb2bd93f465a05ade811cc6cf