[DO NOT MERGE] Turn off carrier privileges grace period feature
Carrier privileges grace period feature was introduced in T
to benefit the switching between the same carrier.

On the cases when switching to different carriers, the feature
introduce a small time window to allow the previous carrier
app to have carrier privileges on upcoming carrier. This has
security concern.

In bug 229418673, Tycho can set a non-Fi pSIM subscription
as the Fi's secondary sub in a sub group, causing confusing
behavior for the end user.

The feature is turn off by default. When a solution that can
fix the security concern has been found, it will be turn on.

Bug: 229418673
Test: atest CarrierPrivilegesTrackerTest
Test: QA test on SS -> DS switching
Test: Telephony sanity(activation, calling, sms, data, switching...)
Change-Id: Ib98a6a97cc590a866a50029237da80634cec2ce3