Fixed delay data switching issue
When allow data on voice call's feature is turned on,
it took a long time for new data to establish on
the voice sub. That's because we didn't tear down
the original network on default data sub.

When temp switching happens, we should tear down the
data network on the original data sub to let new
network setup on the voice sub.

Also when switching happens, we should dynamically update
the network score, so connectivity service will let
the new network created instead of treating the new one
as identical one and then tearing it down.

Fix: 229353836
Test: atest DataNetworkTest + Basic testing + voice call
data temp switching

Change-Id: I026a4d71a737f9a74a588ba236c0fa490054b030