CPT: Fix a missing break statement
When handling ACTION_UICC_ACCESS_RULES_LOADED in the switch
block, the break statement is missing. This will not
impact the funtionality (fall through to default and break)
but introduces a confusing radio log which belongs to
default case.

The change also uses SubscriptionInfoUpdater.simStateString
to print more readable SIM state in logging.

Bug: 227809916
Test: atest CarrierPrivilegesTrackerTest
Test: Telephony sanity(Activation, call, sms, data, switching)
Change-Id: I7d24102a688da705504ab03b44196f49da3401b8