Fix Sync preferred apn and preferred apn set id
When APN db is reset to default, mPreferredSetId is set to 0 along with
the preferred db is reset to empty. However, we do have preferred
profile that's loaded from carrier config, and the preferred db should
be in sync with the mPreferredDataProfile.

Bug: 232890963
Bug: 234725522
Test: manually verified SIM unlock reboot + checked log
Change-Id: I035bd604391ff2aa684ecca0355be9bc79248204
Merged-In: I035bd604391ff2aa684ecca0355be9bc79248204