Correctly apply subId changes in ImsServiceController when SIMs change
In DSDS mode, we were incorrectly calculating which slot to apply subId
changes to when slot 0 had no SIM card and slot 1 moved from having a
SIM card to not having a sim card. In this condition, we would never
clear the sub ID of slot 1, causing a mismatch until re-evaluation of
subscriptions occurred.

This fix corrects this bug by taking into account the condition where
multiple sims have a subId of -1.

Bug: 234562356
Test: atest FrameworksTelephonyTests:ImsServiceControllerTest
Change-Id: I828c4fc59024e84acd5e0be6c4059836aabac6ff