Fix issue where remotely disconnected calls show up as missed.
A recent change was made in ag/17877166 to correct the fact that missed
and rejected calls were being miscategorized.  The fix made there
made the assumption that if a call was LOCALLY disconected (ie the user
pressed disconnect) that it was rejected, otherwise it was missed.

That change failed to take into account the fact that in multiendpoint
scenarios where a call is rejected on an twinned device,
ImsReasonInfo#CODE_REJECTED_ELSEWHERE or
ImsReasonInfo#CODE_REMOTE_CALL_DECLINE can be raised when the call is
rejected on the remote device.

To fix this, we assume for incoming calls that if the call was locally
disconnected OR is the IMS stack reports an ImsReasonInfo code which
maps to DisconnectCause.INCOMING_REJECTED, we'll treat the resulting
disconnet cause as INCOMING_REJECTED.

Fixes: 235126378
Test: Manual test on Verizon network with number twinned into Verizon
Messages app.  Called test number and verified both DUT and the tablet
Verizon app are ringing.  Rejected the call on the tablet.  Verified that
the call shows as "rejected" in the call log on the phone and that the
Telecom disconnect cause is rejected.
Test: ImsPhoneCallTracker#testRejectedElsewhereIsRejected - verifies that
call disconnect cause is REJECTED when the ImsReasonInfo code is
CODE_REJECTED_ELSEWHERE.
Test: ImsPhoneCallTracker#testRemoteCallDeclineIsRejected- verifies that
call disconnect cause is REJECTED when the ImsReasonInfo code is
CODE_REMOTE_CALL_DECLINE.

Change-Id: I9deebe96e4c1c91e00d12fc5dec62c0e5b89afc7