Send IA to modem when SIM switch or refresh
According to b 213517310#comment15, modem wants FWK to send IA again on
those 4 situations. Currently resend IA when SIM switch and refresh are not guaranteed
because in the current logic we only send IA if IA changes.

According to b 227579876, this change should be reverted for U, because
in U modem should only clear IA when FWK requests it to do so.

Test: atest + manual veirify
Fix: 235449106
Change-Id: Ie5c6b2aa8a220f656a70cba5a880cfe4dacf2721